requirements.txt
flask
torch
Pillow
google-generativeai